refactor: #99 드래그 동작 싱글턴 클래스 구현
