+++ BeforeGoing/Presentation/Feature/Scenario/ViewController/ManageScenarioViewController.swift
@@ -50,7 +50,7 @@ final class ManageScenarioViewController: BaseViewController {
         rootView.templateTableView.do {
             $0.delegate = self
             $0.dataSource = self
-            $0.dragDelegate = self
+            $0.dragDelegate = TableViewDragHandler.shared
             $0.dropDelegate = self
             $0.register(ManageScenarioCell.self, forCellReuseIdentifier: ManageScenarioCell.identifier)
         }
@@ -194,17 +194,6 @@ extension ManageScenarioViewController: UITableViewDataSource {
         let config = UISwipeActionsConfiguration(actions: [deleteAction])
         config.performsFirstActionWithFullSwipe = false
         return config
-    }
-}
-
-extension ManageScenarioViewController: UITableViewDragDelegate, DragAction {
-    
-    func tableView(
-        _ tableView: UITableView,
-        itemsForBeginning session: any UIDragSession,
-        at indexPath: IndexPath
-    ) -> [UIDragItem] {
-        return makeDragItems()
     }
 }
